db,scheduler: refactor: eval finalization moves into gradient-db and fires from the effects emitter
